+++ common/exceptions/Cargo.toml
@@ -1,0 +1,24 @@
+[package]
+name = "common-exceptions"
+version = "0.1.0"
+edition = "2021"
+
+[lib]
+doctest = false
+test = false
+
+[dependencies]
+arrow2 = { version = "0.12.0", features = ["io_csv"]}
+futures = "0.3"
+log = "0.4.0"
+paste = "1.0.7"
+serde_json = "1.0.82"
+thiserror = "1.0.31"
+tokio = { version = "1.19.2", features = ["full"] }
+web3 = "0.18.0"
+
+tracing = "0.1"
+opentelemetry = "0.17"
+
+
+[dev-dependencies]
