+++ common/exceptions/src/exception.rs
