+++ ethetl/Cargo.toml
@@ -1,0 +1,36 @@
+[package]
+name = "ethetl"
+version = "0.1.0"
+edition = "2021"
+
+[lib]
+doctest = false
+test = false
+
+[[bin]]
+name = "ethetl"
+path = "bin/ethetl.rs"
+doctest = false
+test = false
+
+[dependencies]
+# Workspace dependencies
+common-exceptions = { path = "../common/exceptions" }
+
+arrow2 = { version = "0.12.0", features = ["io_csv"]}
+clap = { version = "3.2.5", features = ["derive", "env"] }
+env_logger = "0.9.0"
+futures = "0.3"
+log = "0.4.0"
+paste = "1.0.7"
+percentage-rs = "0.1.6"
+serde_json = "1.0.82"
+thiserror = "1.0.31"
+tokio = { version = "1.19.2", features = ["full"] }
+web3 = "0.18.0"
+
+tracing = "0.1"
+opentelemetry = "0.17"
+
+[dev-dependencies]
+goldenfile = "1.3.0"
