+++ ethetl/src/eth/receipts.rs
@@ -12,12 +12,12 @@
 // See the License for the specific language governing permissions and
 // limitations under the License.
 
+use common_exceptions::ErrorCode;
+use common_exceptions::Result;
 use web3::types::TransactionReceipt;
 use web3::types::H256;
 
 use crate::contexts::ContextRef;
-use crate::exceptions::Result;
-use crate::ErrorCode;
 
 pub struct ReceiptFetcher {
     ctx: ContextRef,
