+++ ethetl/src/workers/blocks.rs
@@ -12,12 +12,12 @@
 // See the License for the specific language governing permissions and
 // limitations under the License.
 
+use common_exceptions::Result;
 use futures::stream;
 use futures::stream::StreamExt;
 use futures::stream::TryStreamExt;
 
 use crate::contexts::ContextRef;
-use crate::exceptions::Result;
 use crate::exporters::BlockExporter;
 
 pub struct BlockWorker {
