+++ ethetl/tests/it/fetchers/blocks.rs
@@ -14,9 +14,9 @@
 
 use std::io::Write;
 
+use common_exceptions::Result;
+use ethetl::eth::BlockFetcher;
 use goldenfile::Mint;
-use mars::eth::BlockFetcher;
-use mars::Result;
 
 use crate::common::create_config;
 use crate::common::create_ctx;
